sprint8/visao-computacional/requirements.txt
boto3
python-dotenv
pytz
requests